create the standalone jsfile for checkbox
